+++ CHANGELOG.md
@@ -6,4 +6,6 @@
 - Added /itemInfo to view item's ID and extra info that might be helpful
 - Disabled Vivecraft's Climbing Claws and Jumping boots by default
 - Reorganized config menu with toggles for all new options
-- Added immersiveMC projectile bypass (kinda). Enables items that immersiveMC compatibility blocks in case users disabled immersiveMC's throwing and would rather use this mod's
+- Added immersiveMC projectile bypass (kinda). Enables items 
+that immersiveMC compatibility blocks in case users disabled immersiveMC's throwing and would rather use this mod's
+- Added "Throw Conflicting Items" option to give a way to throw things like bows/crossbows/climbing claws